test(e2e): migrate remaining Detox Snap smokes to Appium
Add remaining Appium Snap smoke specs and remove their Detox twins. Specs-only follow-up to the helpers/PO foundation already on main.
