--- tests/smoke/snaps/test-snap-get-file.spec.ts
@@ -1,54 +1,0 @@
-import { SmokeSnaps } from '../../tags';
-import { loginToApp } from '../../flows/wallet.flow';
-import { navigateToBrowserView } from '../../flows/browser.flow';
-import FixtureBuilder from '../../framework/fixtures/FixtureBuilder';
-import { withFixtures } from '../../framework/fixtures/FixtureHelper';
-import TestSnaps from '../../page-objects/Browser/TestSnaps';
-
-jest.setTimeout(150_000);
-
-describe(SmokeSnaps('Get File Snap Tests'), () => {
-  it('can connect to the get File Snap', async () => {
-    await withFixtures(
-      {
-        fixture: new FixtureBuilder().build(),
-        restartDevice: true,
-        skipReactNativeReload: true,
-        disableSynchronization: true,
-      },
-      async () => {
-        await loginToApp();
-        await navigateToBrowserView();
-        await TestSnaps.navigateToTestSnap();
-
-        await TestSnaps.installSnap('connectGetFileButton');
-      },
-    );
-  });
-
-  it('returns responses for different formats', async () => {
-    await withFixtures(
-      {
-        fixture: new FixtureBuilder().build(),
-        skipReactNativeReload: true,
-        disableSynchronization: true,
-      },
-      async () => {
-        await TestSnaps.tapButton('sendGetFileTextButton');
-        await TestSnaps.checkResultJson('fileResultSpan', { foo: 'bar' });
-
-        await TestSnaps.tapButton('sendGetFileBase64Button');
-        await TestSnaps.checkResultSpan(
-          'fileResultSpan',
-          '"ewogICJmb28iOiAiYmFyIgp9Cg=="',
-        );
-
-        await TestSnaps.tapButton('sendGetFileHexButton');
-        await TestSnaps.checkResultSpan(
-          'fileResultSpan',
-          '"0x7b0a202022666f6f223a2022626172220a7d0a"',
-        );
-      },
-    );
-  });
-});
