--- tests/smoke/snaps/test-snap-jsx.spec.ts
@@ -1,49 +1,0 @@
-import { SmokeSnaps } from '../../tags';
-import { loginToApp } from '../../flows/wallet.flow';
-import { navigateToBrowserView } from '../../flows/browser.flow';
-import FixtureBuilder from '../../framework/fixtures/FixtureBuilder';
-import { withFixtures } from '../../framework/fixtures/FixtureHelper';
-import TestSnaps from '../../page-objects/Browser/TestSnaps';
-import { Assertions, Gestures, Matchers } from '../../framework';
-
-jest.setTimeout(150_000);
-
-describe(SmokeSnaps('JSX Snap Tests'), () => {
-  it('can connect to the JSX Snap', async () => {
-    await withFixtures(
-      {
-        fixture: new FixtureBuilder().build(),
-        restartDevice: true,
-        skipReactNativeReload: true,
-        disableSynchronization: true,
-      },
-      async () => {
-        await loginToApp();
-        await navigateToBrowserView();
-        await TestSnaps.navigateToTestSnap();
-
-        await TestSnaps.installSnap('connectJsx');
-      },
-    );
-  });
-
-  it('displays a modifiable JSX interface', async () => {
-    await withFixtures(
-      {
-        fixture: new FixtureBuilder().build(),
-        skipReactNativeReload: true,
-        disableSynchronization: true,
-      },
-      async () => {
-        await TestSnaps.tapButton('displayJsxButton');
-
-        await Assertions.expectTextDisplayed('0');
-
-        const dynamicButton = Matchers.getElementByText('Increment');
-        await Gestures.tap(dynamicButton);
-
-        await Assertions.expectTextDisplayed('1');
-      },
-    );
-  });
-});
